+++ Database/DatabseSetup.cs
@@ -13,7 +13,7 @@ class DatabaseSetup
     }
     public void CreateComputerTable()
     {
-        var connection = new SqliteConnection("Data Source=database.db");
+        var connection = new SqliteConnection(databaseConfig.ConnectionString);
         connection.Open();
         var command = connection.CreateCommand();
         command.CommandText = @"
