+++ Program.cs
@@ -1,9 +1,12 @@
 ﻿using LabManager.Database;
+using LabManager.Repositories;
 using Microsoft.Data.Sqlite;
 
 
 var databaseConfig = new DatabaseConfig();
 var DatabaseSetup= new DatabaseSetup(databaseConfig);
+
+var computerRepository = new ComputerRepository(databaseConfig);
 
 // Roteamento
 var modelName = args[0];
@@ -13,21 +16,11 @@ if(modelName == "Computer")
 {
     if(modelAction == "List")
     {
-        var connection = new SqliteConnection("Data Source=database.db");
-        connection.Open(); 
-         
-        var command = connection.CreateCommand();
-        command.CommandText = "SELECT * FROM Computers;";
-
-        var reader = command.ExecuteReader();
         Console.WriteLine("Computer List");
-        while(reader.Read())
+        foreach (var computer in computerRepository.GetAll())
         {
-            Console.WriteLine("{0}, {1}, {2} ", reader.GetInt32(0), reader.GetString(1), reader.GetString(2)) ;
+            Console.WriteLine("{0}, {1}, {2}", computer.Id, computer.Ram, computer.Processador);
         }
-
-        reader.Close();
-        connection.Close();
     }
 
     if(modelAction == "New")
@@ -55,20 +48,7 @@ if(modelName == "Lab")
 {
     if(modelAction == "List")
     {
-        var connection = new SqliteConnection("Data Source=database.db");
-        connection.Open();  
-        var command = connection.CreateCommand();
-        command.CommandText = "SELECT * FROM Lab;";
-
-        var reader = command.ExecuteReader();
-        Console.WriteLine("Lab List");
-        while(reader.Read())
-        {
-            Console.WriteLine("{0}, {1}, {2} {3}", reader.GetInt32(0), reader.GetString(1),  reader.GetString(2), reader.GetString(3)) ;
-        }
-
-        reader.Close();
-        connection.Close();
+        
     }
 
     if(modelAction == "New")
